docker/scraper/requirements.txt
pymongo==3.10.1
beautifulsoup4==4.9.3
lxml==4.6.2
python-dotenv==0.15.0
stem==1.8.0
requests==2.24.0
PySocks==1.7.1